Refactored how the spell system tracked each spell
Previously each spell/rune type was tracked with a char that related to what type is was i.e. 'F' for fire, 'W' for water.

I replaced that with an enum and refactored all the methods that previously used chars for their functions. I did this because using a chars for each function would be difficult to understand if I took a prolonged break and for outsiders reading my code. It is far more clear what is going on in the code now.