Finished the Animator Controller for the new Player Character
The animator controller for the new player character is finished. It requires that the fire layer of animations be on top as the fire animations use methods during their animation and as such the fire layer needs to be the base layer. The other animations override the visuals of the fire animation when needed to provide visual clarity of what rune the player has selected or is using.