Refactor of runeSelector and manaBar Script
Created a manaBar script which will handle the character's mana. Will be implemented into the other scripts at a later time.

The runeSelector script had methods and functionality beyond the scope of a rune selector so any methods that was related to the player (i.e. launchSpell(), attack logic, status bools) was moved to the playerScript. Small changes were made to the remaining runeSelector methods to keep the code running.